Moved class to separate file
