+++ ArdalisRating/JsonPolicySerializer.cs
@@ -1,0 +1,14 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace ArdalisRating
+{
+    public class JsonPolicySerializer
+    {
+        public Policy GetPolicyFromJsonString(string jsonString)
+        {
+            return JsonConvert.DeserializeObject<Policy>(jsonString,
+                new StringEnumConverter());
+        }
+    }
+}
